requirements.txt
# Irrigation Intelligente - Mauritanie
# Dépendances Python

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Big Data / Streaming
kafka-python>=2.0.0
pyspark>=3.5.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0

# Utilities
python-dotenv>=1.0.0